refactor: 패키지 구조 변경
- 내부 서비스와 출력 기능을 구분하기 위해 service 패키지 생성
