+++ src/main/java/lotto/service/LottoPurchaseService.java
@@ -1,7 +1,8 @@
-package lotto;
+package lotto.service;
 
 import java.util.ArrayList;
 import java.util.List;
+import lotto.Lotto;
 
 public class LottoPurchaseService {
