+++ src/main/java/lotto/service/Standard.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.service;
 
 import java.util.List;
