+++ src/test/java/lotto/WinningHistoryTest.java
@@ -3,6 +3,8 @@ package lotto;
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
+import lotto.service.LottoRank;
+import lotto.service.WinningHistory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
